+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 17.0
 + **Breaking:** Replaced the entire context API.
++ **Breaking:** Rename `isolate` to `teardownOnError`.
 
 ## 16.2
 + Allow `undefined` in context values.